requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
mercadopago==2.2.1
django-cors-headers==4.3.1
django-filter==24.1
gunicorn==21.2.0
whitenoise==6.6.0
